Moved files into util folder to tidy up.
Built the .npmignore, updated package.json, created a new tkyodrift.js file to handle CLI parsing.

old tkyodrift.js file updated to reflect it's purpose, it is now oneOffEmb.js in util.
