+++ tkyo-drift/.npmignore
@@ -1,4 +1,3 @@
 
 __pycache__/
 data/
-test/
