--- tkyo-drift/index.js
@@ -1,4 +1,0 @@
-import tkyoDrift from './tkyoDrift.js';
-import tkyoDriftSetTrainingHook from './tkyoDriftSetTrainingHook.js';
-
-export { tkyoDrift, tkyoDriftSetTrainingHook };
