+++ tkyo-drift/package.json
@@ -1,25 +1,45 @@
 {
-  "name": "tkyodrifttest1",
+  "name": "tkyodrift",
   "version": "1.0.0",
-  "description": "Test AI Drift and Embed Testing Data",
-  "main": "index.js",
-  "scripts": {
-    "test": "test.js"
+  "description": "Lightweight CLI tool and library for detecting AI model drift using embeddings and scalar metrics. Tracks semantic, conceptual, and lexical change over time.",
+  "main": "./tkyoDrift.js",
+  "bin":{
+    "tkyo": "./tkyoDrift.js"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/oslabs-beta/tkyo-drift"
   },
-  "keywords": [],
+  "keywords": [
+    "drift-detection",
+    "ai-monitoring",
+    "embedding",
+    "model-drift",
+    "semantic-drift",
+    "concept-drift",
+    "lexical-drift",
+    "ai-evaluation",
+    "machine-learning",
+    "transformers",
+    "xenova",
+    "huggingface",
+    "vector-embedding",
+    "cosine-similarity",
+    "euclidean-distance",
+    "scalars",
+    "population-stability-index",
+    "monitoring",
+    "model-evaluation",
+    "cli-tool"
+  ],
   "authors": "Anthony Heinz, Wing Man Lam, Monique Nunez, Chris Antico, Milti Symeonoglou",
-  "license": "ISC",
+  "license": "MIT",
   "dependencies": {
     "@xenova/transformers": "^2.17.2",
     "chalk": "^5.4.1",
     "cli-table3": "^0.6.5",
     "fs": "^0.0.1-security",
     "path": "^0.12.7",
-    "tkyodrifttest1": "^1.0.0",
     "uuid": "^11.1.0"
   },
   "type": "module"
