+++ tkyo-drift/util/makeLogEntry.js
@@ -1,6 +1,6 @@
 import fs from 'fs';
 import path from 'path';
-import { OUTPUT_DIR } from '../tkyoDrift.js';
+import { OUTPUT_DIR } from './oneOffEmb.js';
 
 export default function makeLogEntry(id, mathObject, type) {
   let logPath = '';
